requirements.txt
asgiref==3.8.1
attrs==25.1.0
distlib==0.3.9
Django==5.1.6
django-filter==25.1
djangorestframework==3.15.2
djangorestframework_simplejwt==5.4.0
drf-spectacular==0.28.0
drf-spectacular-sidecar==2025.2.1
filelock==3.17.0
inflection==0.5.1
jsonschema==4.23.0
jsonschema-specifications==2024.10.1
pbr==6.1.1
platformdirs==4.3.6
psycopg2==2.9.10
PyJWT==2.10.1
PyYAML==6.0.2
referencing==0.36.2
rpds-py==0.23.1
setuptools==75.8.0
sqlparse==0.5.3
stevedore==5.4.0
typing_extensions==4.12.2
tzdata==2025.1
uritemplate==4.1.1
virtualenv==20.29.2
virtualenv-clone==0.5.7
virtualenvwrapper-win==1.2.7
